encapsulated Request<Void> properly
