+++ checkstyle.xml
@@ -237,7 +237,6 @@
             <property name="allowedAbbreviations" value="IT"/>
             <property name="allowedAbbreviationLength" value="1"/>
         </module>
-        <module name="AbstractClassName"/>
         <module name="ClassTypeParameterName"/>
         <module name="ConstantName"/>
         <module name="LocalFinalVariableName">
