+++ src/main/java/com/amihaiemil/charles/aws/AmazonEsRepository.java
@@ -39,6 +39,13 @@ import com.amazonaws.AmazonServiceException;
 import com.amazonaws.http.HttpResponse;
 import com.amihaiemil.charles.DataExportException;
 import com.amihaiemil.charles.WebPage;
+import com.amihaiemil.charles.aws.requests.AwsDelete;
+import com.amihaiemil.charles.aws.requests.AwsHead;
+import com.amihaiemil.charles.aws.requests.AwsHttpHeaders;
+import com.amihaiemil.charles.aws.requests.AwsHttpRequest;
+import com.amihaiemil.charles.aws.requests.AwsPost;
+import com.amihaiemil.charles.aws.requests.EsHttpRequest;
+import com.amihaiemil.charles.aws.requests.SignedRequest;
 import com.amihaiemil.charles.github.AwsEsRepository;
 
 /**
