+++ src/main/java/com/amihaiemil/charles/aws/requests/AwsDelete.java
@@ -23,7 +23,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package com.amihaiemil.charles.aws;
+package com.amihaiemil.charles.aws.requests;
 
 import com.amazonaws.Request;
 import com.amazonaws.http.HttpMethodName;
@@ -35,7 +35,7 @@ import com.amazonaws.http.HttpMethodName;
  * @since 1.0.0
  *
  */
-public final class AwsDelete<T> implements AwsHttpRequest<T> {
+public final class AwsDelete<T> extends AwsHttpRequest<T> {
 
     /**
      * Base request.
@@ -57,7 +57,7 @@ public final class AwsDelete<T> implements AwsHttpRequest<T> {
     }
 
     @Override
-    public Request<Void> request() {
+    Request<Void> request() {
         return this.base.request();
     }
 }
