+++ src/main/java/com/amihaiemil/charles/aws/requests/AwsHttpHeaders.java
@@ -23,22 +23,19 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package com.amihaiemil.charles.aws;
+package com.amihaiemil.charles.aws.requests;
 
-import java.io.InputStream;
-
+import java.util.Map;
 import com.amazonaws.Request;
-import com.amazonaws.http.HttpMethodName;
-
 
 /**
- * Http POST request sent to AWS.
+ * Aws HTTP request with set headers.
  * @author Mihai Andronache (amihaiemil@gmail.com)
  * @version $Id$
  * @since 1.0.0
  *
  */
-public final class AwsPost<T> implements AwsHttpRequest<T> {
+public final class AwsHttpHeaders<T> extends AwsHttpRequest<T> {
 
     /**
      * Base request.
@@ -48,13 +45,12 @@ public final class AwsPost<T> implements AwsHttpRequest<T> {
     /**
      * Ctor.
      * @param req Base AwsHttpRequest.
-     * @param content InputStream containing this request's content.
+     * @param headers Map of headers to set on the request. Key - headername, value - its value.
      */
-    public AwsPost(AwsHttpRequest<T> req, InputStream content) {
-    	this.base = req;
-        this.base.request().setHttpMethod(HttpMethodName.POST);
-    	this.base.request().setContent(content);
-	}
+    public AwsHttpHeaders(AwsHttpRequest<T> req, Map<String, String> headers) {
+        this.base = req;
+        this.base.request().getHeaders().putAll(headers);
+    }
 
     @Override
     public T perform() {
@@ -62,7 +58,7 @@ public final class AwsPost<T> implements AwsHttpRequest<T> {
     }
 
     @Override
-    public Request<Void> request() {
-    	return this.base.request();
+    Request<Void> request() {
+        return this.base.request();
     }
 }
